+++ packages/core/src/index.ts
@@ -16,3 +16,5 @@ export * from './scripts/migrate-oz-cli-project';
 
 export { logWarning } from './utils/log';
 export { setProxyKind } from './set-proxy-kind';
+
+export { UpgradeableContract } from './standalone';