+++ packages/core/src/validate/index.ts
@@ -13,6 +13,7 @@ export {
   isUpgradeSafe,
   inferProxyKind,
 } from './query';
+export { UpgradeableContractErrorReport } from './report';
 
 // Backwards compatibility
 export { silenceWarnings } from '../utils/log';