+++ packages/core/src/validate/query.ts
@@ -13,8 +13,7 @@ export function assertUpgradeSafe(data: ValidationData, version: Version, opts: 
   const dataV3 = normalizeValidationData(data);
   const [contractName] = getContractNameAndRunValidation(dataV3, version);
 
-  let errors = getErrors(dataV3, version);
-  errors = processExceptions(contractName, errors, opts);
+  const errors = getErrors(dataV3, version, opts);
 
   if (errors.length > 0) {
     throw new ValidationErrors(contractName, errors);
@@ -103,7 +102,7 @@ export function getUnlinkedBytecode(data: ValidationData, bytecode: string): str
   return bytecode;
 }
 
-export function getErrors(data: ValidationData, version: Version): ValidationError[] {
+export function getErrors(data: ValidationData, version: Version, opts: ValidationOptions = {}): ValidationError[] {
   const dataV3 = normalizeValidationData(data);
   const [contractName, runValidation] = getContractNameAndRunValidation(dataV3, version);
   const c = runValidation[contractName];
@@ -119,7 +118,7 @@ export function getErrors(data: ValidationData, version: Version): ValidationErr
     });
   }
 
-  return errors;
+  return processExceptions(contractName, errors, opts);
 }
 
 function getAllMethods(runValidation: ValidationRunData, contractName: string): string[] {